% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{createDataTableFromNCDF}
\alias{createDataTableFromNCDF}
\title{Creates a data.table from the ncdf4 input
Not for external use}
\usage{
createDataTableFromNCDF(variable, validValues)
}
\description{
Creates a data.table from the ncdf4 input
Not for external use
}

