% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eobsR.R
\docType{package}
\name{eobsR}
\alias{eobsR}
\alias{eobsR-package}
\title{eobsR: A package to import and visualize the EOBS data}
\description{
The package provides two functions:
importEOBS and runExample.
}
\section{eobsR functions}{

Do the following
}

