% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{getOpenDapValues}
\alias{getOpenDapValues}
\title{Accesses the OPeNDAB server}
\usage{
getOpenDapValues(opendapURL, variableName, bbox, period)
}
\arguments{
\item{opendapURL}{String}

\item{variableName}{String which variable to get}
}
\description{
Accesses the OPeNDAB server
}
\note{
This function is based on the script by Maarten Plieger
https://publicwiki.deltares.nl/display/OET/OPeNDAP+subsetting+with+R
}

