% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{importEOBS.bbox}
\alias{importEOBS.bbox}
\title{Cuts a bbox from the ncdf file and returns the values as data.table
Not for external use}
\usage{
importEOBS.bbox(variable, period, bbox, grid)
}
\description{
Cuts a bbox from the ncdf file and returns the values as data.table
Not for external use
}

