% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{importEOBS}
\alias{importEOBS}
\title{Imports EOBS data}
\usage{
importEOBS(variable, period, area, grid, na.rm = TRUE, download = TRUE)
}
\arguments{
\item{variable}{String from ('tg', 'tn', 'tx, ...) / TODO: vector}

\item{period}{Temporal object (should be only one temporal block)}

\item{area}{Spatial object}

\item{grid}{String from ('0.25reg', '0.5reg', '0.25rot', '0.5rot')}

\item{download}{Boolean indicating whether to download}

\item{removeNA}{Boolean indicating if rows with NA can be deleted / TODO: This has to be fine tuned}
}
\description{
Imports EOBS data
}
\note{
Fixed file is loaded instead of OPeNDAP request, i.e. variable,
 period, and grid are deprecated
}

