% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{LocalImportEOBS}
\alias{LocalImportEOBS}
\title{Import EOBS data from local file}
\usage{
LocalImportEOBS(variable, filename, period = NULL, area = NULL,
  na.rm = TRUE)
}
\arguments{
\item{variable}{String from ('tg', 'tn', 'tx, ...)}

\item{filename}{String containing the path to the ncdf file}

\item{period}{Either numeric, timeBased or ISO-8601 style 
(see \code{\link[xts]{.subset.xts}})}

\item{area}{Either SpatialPolygons or SpatialPolygonsDataFrame object (see
package sp)}

\item{na.rm}{Boolean indicating if rows with NA can be deleted}
}
\description{
Import EOBS data from local file
}
\note{
Should be merged with \code{\link{importEOBS}}
}

