% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{removeOutsiders}
\alias{removeOutsiders}
\title{Removes points outside of the SpatialPolygons
Not for external use}
\usage{
removeOutsiders(data, area)
}
\description{
Removes points outside of the SpatialPolygons
Not for external use
}

