% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{SanitizeInput}
\alias{SanitizeInput}
\title{Checks whether the input to importEOBS is valid or not}
\usage{
SanitizeInput(variable, period, area, grid)
}
\description{
Checks whether the input to importEOBS is valid or not
}

