% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{specifyURL}
\alias{specifyURL}
\title{Specifies the url based on the variableName and the grid}
\usage{
specifyURL(variableName, grid)
}
\description{
Specifies the url based on the variableName and the grid
}

