% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{importEOBS}
\alias{importEOBS}
\title{Imports EOBS data}
\usage{
importEOBS(variable, period, area, grid, removeNA = TRUE, download = TRUE)
}
\arguments{
\item{variable}{String from ('TG', 'TN', 'TX, ...) / TODO: vector}

\item{period}{Temporal object}

\item{area}{Spatial object}

\item{grid}{String from ('0.25', '0.5', 'R0.25', ...)}

\item{removeNA}{Boolean indicating if rows with NA can be deleted / TODO: This has to be fine tuned}

\item{download}{Boolean indicating whether to download}
}
\description{
Imports EOBS data
}
\note{
Fixed file is loaded instead of OPeNDAP request, i.e. variable,
 period, and grid are deprecated
}

