% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEOBS.R
\name{importEOBS}
\alias{importEOBS}
\title{Imports EOBS data}
\usage{
importEOBS(variable, period, area, grid, na.rm = TRUE, download = TRUE)
}
\arguments{
\item{variable}{String from ('tg', 'tn', 'tx, ...)}

\item{period}{Either numeric, timeBased or ISO-8601 style 
(see \code{\link[xts]{.subset.xts}})}

\item{area}{Either SpatialPolygons or SpatialPolygonsDataFrame object (see
package sp)}

\item{grid}{String from ('0.25reg', '0.50reg', '0.25rot', '0.50rot')}

\item{na.rm}{Boolean indicating if rows with NA can be deleted}

\item{download}{Boolean indicating whether to download}
}
\description{
Imports EOBS data
}

